+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Change posts more text customizer control label to 'More link text'
 - Rename `languages/` directory to `lang/`
 - Move reusable methods in `Setups\MetaBoxes\Layout` to `Setups\MetaBoxes\AbstractMetaBox`
-- Prefix registered thumbnail sizes with `jentil-`
+- Prefix registered thumbnail sizes, widget and sidebar ids with `jentil-`
 - Append enqueued assets URL with last modified time for cache busting
 - Split up thumbnail setup into separate setups